Inhibit multishift toggle when disabling it is unsafe
If multiple shifts of the same course/type are selected, and the
multi-shift ability is disabled, we get weird behavior when manipulating
those shifts.

Bumps typescript target to ES2015: it's modern enough, and means no
for-of transpilation (which is behind a toggle for es5 anyway). It's
extremely unlikely that anyone will use a browser old enough to cause
problems with this.